Filesystem

+++ src/system/hal/common/hal.c
@@ -7,13 +7,13 @@
 
 #include "hal.h"
 
-uint32_t hal_get_address_value(uint32_t port, uint8_t offset)
+uint32_t hal_get_address_value(uint32_t port, uint32_t offset)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	return *reg;
 }
 
-uint32_t hal_bitmask_write(uint32_t port, uint8_t offset, uint32_t mask, uint8_t size)
+uint32_t hal_bitmask_write(uint32_t port, uint32_t offset, uint32_t mask, uint8_t size)
 {
 	if(size == 32) {
 		mmio_t reg = hal_get_register(port, offset);
@@ -40,14 +40,14 @@ uint32_t hal_bitmask_write(uint32_t port, uint8_t offset, uint32_t mask, uint8_t
 	return 0;
 }
 
-void hal_bitmask_set(uint32_t port, uint8_t offset, uint32_t mask)
+void hal_bitmask_set(uint32_t port, uint32_t offset, uint32_t mask)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	*reg |= mask;
 }
 
 
-void hal_bitmask_clear(uint32_t port, uint8_t offset, uint32_t mask)
+void hal_bitmask_clear(uint32_t port, uint32_t offset, uint32_t mask)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	*reg &= ~mask;
